DDF-2367 make GmdConverter and GmdTransformer extendable
